ciniguardai1/requirements.txt
fastapi
uvicorn[standard]
opencv-python
ultralytics
supabase
python-dotenv
python-multipart
websockets
